Revert "WIP package structure and style overhaul"
This reverts commit 87998bd5c4e0b9623448d0928fec911ea843323a.
