--- digipyro/__init__.py
@@ -1,2 +1,0 @@
-"""digipyro: digitally rotate a video in python."""
-from .digipyro import start
