+++ synth.py
